♻️ Trekk logikk for å validere perioder for innvilgelse ut i egen fil for å kunne gjennbruke i V2 av beregning
