+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/barnetilsyn/beregning/TilsynBarnBeregningService.kt
@@ -1,21 +1,16 @@
 package no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning
 
-import no.nav.tilleggsstonader.kontrakter.felles.overlapper
 import no.nav.tilleggsstonader.sak.behandling.domain.Saksbehandling
 import no.nav.tilleggsstonader.sak.felles.domain.BarnId
 import no.nav.tilleggsstonader.sak.felles.domain.BehandlingId
 import no.nav.tilleggsstonader.sak.infrastruktur.database.repository.findByIdOrThrow
-import no.nav.tilleggsstonader.sak.infrastruktur.exception.brukerfeilHvis
-import no.nav.tilleggsstonader.sak.infrastruktur.exception.brukerfeilHvisIkke
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvis
 import no.nav.tilleggsstonader.sak.util.YEAR_MONTH_MIN
 import no.nav.tilleggsstonader.sak.util.datoEllerNesteMandagHvisLørdagEllerSøndag
 import no.nav.tilleggsstonader.sak.util.toYearMonth
 import no.nav.tilleggsstonader.sak.vedtak.TypeVedtak
 import no.nav.tilleggsstonader.sak.vedtak.VedtakRepository
-import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBeregningUtil.brukPerioderFraOgMedRevurderFra
-import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBeregningUtil.brukPerioderFraOgMedRevurderFraMåned
-import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBeregningUtil.erOverlappMellomStønadsperioderOgUtgifter
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBarnBeregningValidering.validerPerioderForInnvilgelse
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBeregningUtil.tilAktiviteterPerMånedPerType
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBeregningUtil.tilDagerPerUke
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBeregningUtil.tilUke
@@ -250,67 +245,4 @@ class TilsynBarnBeregningService(
         vilkårperiodeRepository
             .findByBehandlingIdAndResultat(behandlingId, ResultatVilkårperiode.OPPFYLT)
             .tilAktiviteter()
-
-    private fun validerPerioderForInnvilgelse(
-        stønadsperioder: List<StønadsperiodeBeregningsgrunnlag>,
-        aktiviteter: List<Aktivitet>,
-        utgifter: Map<BarnId, List<UtgiftBeregning>>,
-        typeVedtak: TypeVedtak,
-        revurderFra: LocalDate?,
-    ) {
-        if (typeVedtak == TypeVedtak.OPPHØR) {
-            return
-        }
-        validerStønadsperioder(stønadsperioder)
-        validerAktiviteter(aktiviteter)
-        validerUtgifter(utgifter)
-        validerStønadsperioderOverlapperUtgifterEtterRevurderFra(stønadsperioder, utgifter, revurderFra)
-    }
-
-    private fun validerStønadsperioderOverlapperUtgifterEtterRevurderFra(
-        stønadsperioder: List<StønadsperiodeBeregningsgrunnlag>,
-        utgifter: Map<BarnId, List<UtgiftBeregning>>,
-        revurderFra: LocalDate?,
-    ) {
-        brukerfeilHvisIkke(
-            erOverlappMellomStønadsperioderOgUtgifter(
-                stønadsperioder = stønadsperioder.brukPerioderFraOgMedRevurderFra(revurderFra),
-                utgifter = utgifter.brukPerioderFraOgMedRevurderFraMåned(revurderFra),
-            ),
-        ) {
-            if (revurderFra != null) {
-                "Kan ikke innvilge når det ikke finnes noen overlapp mellom målgruppe, aktivitet og utgifter etter revurder fra dato"
-            } else {
-                "Kan ikke innvilge når det ikke finnes noen overlapp mellom målgruppe, aktivitet og utgifter"
-            }
-        }
-    }
-
-    private fun validerStønadsperioder(stønadsperioder: List<StønadsperiodeBeregningsgrunnlag>) {
-        brukerfeilHvis(stønadsperioder.isEmpty()) {
-            "Kan ikke innvilge når det ikke finnes noen overlappende målgruppe og aktivitet"
-        }
-    }
-
-    private fun validerAktiviteter(aktiviteter: List<Aktivitet>) {
-        feilHvis(aktiviteter.isEmpty()) {
-            "Aktiviteter mangler"
-        }
-    }
-
-    private fun validerUtgifter(utgifter: Map<BarnId, List<UtgiftBeregning>>) {
-        brukerfeilHvis(utgifter.values.flatten().isEmpty()) {
-            "Kan ikke innvilge når det ikke finnes noen utgiftsperioder"
-        }
-        utgifter.entries.forEach { (_, utgifterForBarn) ->
-            feilHvis(utgifterForBarn.overlapper()) {
-                "Utgiftsperioder overlapper"
-            }
-
-            val ikkePositivUtgift = utgifterForBarn.firstOrNull { it.utgift < 0 }?.utgift
-            feilHvis(ikkePositivUtgift != null) {
-                "Utgiftsperioder inneholder ugyldig utgift: $ikkePositivUtgift"
-            }
-        }
-    }
 }
